Invert the default commit behaviour when modifying routes.
Currently it's necessary to specify `:skip_commit => true` to prevent
the api client from committing the routes after each change.  It turns
out that in most cases, we need to use this option because we're
adding more than one route at a time.

This changes the default behaviout to not commit the route, with an
option to enable the commit.